requirements.txt
spacy>=3.7.0
pymupdf>=1.23.0
python-docx>=1.1.0
